major code refactoring for milestone 3.1

+++ Software_Engineering/src/main/java/factory/shared/Position.java
@@ -5,19 +5,19 @@ package factory.shared;
  */
 public class Position {
 
-	public void setPositionX(double xPos) {
+	public void setPositionX(int xPos) {
 		//TODO
 	}
 
-	public void setPositionY(double yPos) {
+	public void setPositionY(int yPos) {
 		//TODO
 	}
 
-	public double getPositionX() {
+	public int getPositionX() {
 		return 0;	//TODO
 	}
 
-	public double getPositionY() {
+	public int getPositionY() {
 		return 0;	//TODO
 	}
 }
